backend/requirements.txt
fastapi
uvicorn
diffusers
transformers
accelerate
safetensors
torch
rembg
onnxruntime
pillow
opencv-python
python-multipart
